Cleaned up example code
